+++ app/services/forest_liana/intercom_conversation_getter.rb
@@ -1,5 +1,5 @@
 module ForestLiana
-  class IntercomConversationGetter
+  class IntercomConversationGetter < IntegrationBaseGetter
     attr_accessor :record
 
     def initialize(params)