+++ app/services/forest_liana/stripe_payments_getter.rb
@@ -1,5 +1,5 @@
 module ForestLiana
-  class StripePaymentsGetter
+  class StripePaymentsGetter < StripeBaseGetter
     attr_accessor :records
 
     def initialize(params, secret_key, reference)
@@ -63,31 +63,6 @@ module ForestLiana
       if pagination? && @params[:ending_before]
         @params[:ending_before]
       end
-    end
-
-    def limit
-      return 10 unless pagination?
-
-      if @params[:page][:size]
-        @params[:page][:size].to_i
-      else
-        10
-      end
-    end
-
-    def pagination?
-      @params[:page]
-    end
-
-    def collection
-      @params[:collection].singularize.camelize.constantize
-    end
-
-    def field
-      ForestLiana.integrations[:stripe][:mapping].select { |value|
-        value.split('.')[0] == ForestLiana::SchemaUtils
-          .find_model_from_collection_name(@params[:collection]).try(:name)
-      }.first.split('.')[1]
     end
   end
 end