+++ app/services/forest_liana/stripe_source_getter.rb
@@ -1,5 +1,5 @@
 module ForestLiana
-  class StripeSourceGetter
+  class StripeSourceGetter < StripeBaseGetter
     attr_accessor :record
 
     def initialize(params, secret_key, reference)
@@ -20,17 +20,6 @@ module ForestLiana
       @record.customer = collection.find_by(query)
 
       @record
-    end
-
-    def collection
-      @params[:collection].singularize.camelize.constantize
-    end
-
-    def field
-      ForestLiana.integrations[:stripe][:mapping].select { |value|
-        value.split('.')[0] == ForestLiana::SchemaUtils
-          .find_model_from_collection_name(@params[:collection]).try(:name)
-      }.first.split('.')[1]
     end
   end
 end