+++ app/services/forest_liana/stripe_subscription_getter.rb
@@ -1,5 +1,5 @@
 module ForestLiana
-  class StripeSubscriptionGetter
+  class StripeSubscriptionGetter < StripeBaseGetter
     attr_accessor :record
 
     def initialize(params, secret_key, reference)
@@ -18,17 +18,6 @@ module ForestLiana
         @record.customer = nil
       end
       @record
-    end
-
-    def collection
-      @params[:collection].singularize.camelize.constantize
-    end
-
-    def field
-      ForestLiana.integrations[:stripe][:mapping].select { |value|
-        value.split('.')[0] == ForestLiana::SchemaUtils
-          .find_model_from_collection_name(@params[:collection]).try(:name)
-      }.first.split('.')[1]
     end
   end
 end