+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### Fixed
 - Stripe Integration - Fix the global integration when mapped on namespaced models.
+- Stripe Integration - Fix the access to customer subscriptions and bank accounts.
 
 ## RELEASE 1.9.7 - 2017-10-30
 ### Fixed